code_gen/requirements.txt
boto3>=1.40.0
botocore>=1.40.0
opensearch-py>=2.0.0
python-dotenv>=1.0.0
requests>=2.32.0
urllib3>=2.5.0
